Add user auth in Incomes and Expenses

+++ app/Http/Controllers/ExpenseController.php
@@ -5,12 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Expense;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ExpenseController extends Controller
 {
     public function index()
     {
-        $expenses = Expense::with('category')->get();
+        $expenses = Expense::with('category') ->where('user_id', Auth::id()) ->get();
         return view('expenses.index', compact('expenses'));
     }
 
@@ -23,25 +24,39 @@ class ExpenseController extends Controller
     public function store(Request $request)
     {
         $request->validate([
+            'user_id' => 'required|exists:users,id',
             'category_id' => 'required|exists:categories,id',
             'amount' => 'required|numeric',
             'description' => 'nullable|string',
             'date' => 'required|date',
         ]);
 
-        Expense::create($request->all());
+        Expense::create(array_merge($request->all(), ['user_id' => Auth::id()]));
 
         return redirect()->route('expenses.index')->with('success', 'Expense created successfully.');
     }
 
+    public function show(Expense $expense)
+    {
+        return view('expense.show', compact('expense'));
+    }
+
     public function edit(Expense $expense)
     {
+        if ($expense->user_id !==  Auth::id()) {
+            abort(403);
+        }
+
         $categories = Category::all();
         return view('expenses.edit', compact('expense', 'categories'));
     }
 
     public function update(Request $request, Expense $expense)
     {
+        if ($expense->user_id !==  Auth::id()) {
+            abort(403);
+        }
+
         $request->validate([
             'category_id' => 'required|exists:categories,id',
             'amount' => 'required|numeric',
@@ -56,6 +71,9 @@ class ExpenseController extends Controller
 
     public function destroy(Expense $expense)
     {
+        if ($expense->user_id !==  Auth::id()) {
+            abort(403);
+        }
         $expense->delete();
         return redirect()->route('expenses.index')->with('success', 'Expense deleted successfully.');
     }
